fix/structure-flow
There is no need to navigate to DetailScreen cause i can show GoogleTab from first fragment as the instructions says.
