+++ app/src/main/java/com/app/hackernews/ui/home/viewModel/HomeScreenViewModel.kt
@@ -1,5 +1,9 @@
 package com.app.hackernews.ui.home.viewModel
 
+import android.content.Intent
+import android.net.Uri
+import androidx.activity.result.ActivityResultLauncher
+import androidx.browser.customtabs.CustomTabsIntent
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.app.hackernews.data.network.models.RequestState
@@ -24,6 +28,20 @@ class HomeScreenViewModel
                 _androidNewsState.value = RequestState.Loading
                 _androidNewsState.value = androidNewsUseCase.invoke()
             }
+
+        fun openCustomTab(
+            url: String,
+            launcher: ActivityResultLauncher<Intent>,
+        ) {
+            val intent = CustomTabsIntent.Builder().build()
+
+            val customTabsIntent =
+                intent.intent.apply {
+                    data = Uri.parse(url)
+                }
+
+            launcher.launch(customTabsIntent)
+        }
 
         init {
             getAndroidNews()
